polymarket_trader/requirements.txt
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
eth-account>=0.11.0
web3>=6.15.0
py-clob-client>=0.14.0
numpy>=1.26.0
colorama>=0.4.6
